+++ db/migrate/20260807000001_add_deleted_at_to_rivet_cms_documents.rb
@@ -1,0 +1,6 @@
+class AddDeletedAtToRivetCmsDocuments < ActiveRecord::Migration[7.2]
+  def change
+    add_column :rivet_cms_documents, :deleted_at, :datetime
+    add_index :rivet_cms_documents, :deleted_at
+  end
+end